MockEth cleanup to merge non mock changes

+++ packages/arb-validator/mockbridge/executionChallenge.go
@@ -217,7 +217,7 @@ func (c *ExecutionChallenge) ChooseSegment(
 ) error {
 	bisectionHashes := make([]common.Hash, 0, len(assertions))
 	for i := range assertions {
-		stepCount := structures.CalculateBisectionStepCount(uint32(i), uint32(len(assertions)), totalSteps)
+		stepCount := structures.CalculateBisectionStepCount(uint64(i), uint64(len(assertions)), totalSteps)
 		bisectionHashes = append(
 			bisectionHashes,
 			structures.ExecutionDataHash(stepCount, preconditions[i].Hash(), assertions[i].Hash()),
